fix: Don't use Status for managing breaker state
The `Status` object is really only for keeping track of various events
that occur within the breaker. The CircuitBreaker doesn't really need to
manage those counts. And it especially doesn't need to depend on them
for managing state.

Move `Status` to its own module, and have it use breaker events to
manage its own stats.