feat: circuits now have a name based on the action
The name of the circuit can be supplied as an option.

```js
const circuit = circuitBreaker(() => {}, { name: 'identity' });
```

If not option is provided, the name will be based on the action's name
if there is one.

```js
const square = (x) => x*x;
const circuit = circuitBreaker(square);
circuit.name // 'square'
```

If no option is provided, and the circuit's action is an anonymous
function, then a UUID is generated.