chore: simplify status.js implementation - add doc
Added some JSDoc text to circuit and status objects. Greatly simplified
how the status object behaves. It no longer has a dependency on the
circuit itself. The circuit sets up all of the listeners for the status
object and increments it as necessary. This makes the status object much
cleaner, and allows status to be focused primarily on managing the
rolling stats window, vs monitoring the circuit for events.