feat: prefer an error percentage threshold
Until now, the circuit breaker could only be configured to trip after a
fixed number of failures. This is problematic, because those failures
may accumulate over a long period of time, causing the circuit to open
once the threshold has been hit - even if the failure percentage is
very, very low.

Now when using `options.maxFailures` a deprecation warning is printed to
the error console. The new option is `errorThresholdPercentage`.

Fixes: https://github.com/bucharest-gold/opossum/issues/37