feat: add circuit.shutdown() to shut a circuit down
Shutting down a circuit will cause it to always return a
rejected promise with an `ESHUTDOWN` `Error.code` when
`circuit.fire()` is called, and the Hystrix stats stop
accumulating. Once a circuit has been shut down, it cannot
be reset or used again.

When using the `shutdown()` method on a circuit, all listeners
are cleaned up internally, preventing memory leaks.

Fixes: https://github.com/nodeshift/opossum/issues/248
Fixes: https://github.com/nodeshift/opossum/issues/181
Fixes: https://github.com/nodeshift/opossum/issues/245