build: don't fail coverage from untrusted forks
The CircleCI tool by default will not expose secrets to untrusted
forks in the ci build. In this case, that means that
`$CODACY_PROJECT_TOKEN` is not set before the build, and the final
step of the CI build is to send Codacy coverage data. This change
prevents this from causing a build failure when the variable is
unset due to a PR from an untrusted fork (i.e. the community).