feat: refactor Prometheus and Hystrix metrics
Instead of continuing to add different metrics modules to the
circuit breaker API, we can extract the `PrometheusMetrics` and
`HystrixStats` classes as their own modules. All they need is a
handle to the circuit itself so they can listen for events and/or
monitor the circuit stats. This allows us to continue maintaining
things like Hystrics that are deprecated upstream, without polluting
our circuit breaker implementation.