bug: fix stuck halfOpen when using errorFilter (#389)
* Fix stuck halfOpen when using errorFilter

When an error is filtered by an errorFilter the request is
considered neither success nor failure, leaving the circuit
in an inconsistent state. This fixes that issue by emitting
success event when the error is filtered.

* Call fallback when an error is filtered

When an error is filtered, it is considered a success, however we need a
result to resolve with, this way, we should call fallback to get a valid
result. If there is no fallback, we reject to give the application a
chance to handle the error.