chore: update CI to include build+test:headless (#680)
* chore: update CI to include build+test:headless

Because some of the dependabot / snyk PRs could update dependencies like
webpack and friends, these checks should be included for pull requests.

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: add NODE_OPTIONS=--openssl-legacy-provider

Signed-off-by: Lance Ball <lball@redhat.com>

* chore: update CI to only build for current

This commit modifies CI so that the build matrix only uses the two
previous releases, and runs a separate action for `current` (18x at the
moment). This is because some things changed in Node.js 18 and I need
to set an environment variable when building to avoid error. However,
setting this environment variable results in an error for older versions
of Node.js.

Signed-off-by: Lance Ball <lball@redhat.com>

Signed-off-by: Lance Ball <lball@redhat.com>