feat: add option for user-provided EventEmitter to control buckets (#809)
The motivation for this PR. I added an option for a user-provided EventEmitter that would handle the rotating of the buckets. The idea here is that you could do something like this in your app
```
const bucketController = new EventEmitter();
setInterval(() => {
   bucketController.emit('rotate')
}, 1000)
```
And then pass `bucketController` into opossum. Opossum would listen for that event and then rotate the buckets.


---------

Co-authored-by: Gautam Jethwani <gjethwani@atlassian.com>